Pooling for Online Storage Basic Version works, renamed some stuff to better fit convention, removed some old code

+++ RandomForestWithGPs/src/Base/Observer.cc
@@ -44,5 +44,5 @@ void Subject::notify(const unsigned int event){
 }
 
 unsigned int Subject::numberOfObservers() const{
-	return m_observers.size();
+	return (unsigned int) m_observers.size();
 }
